requirements.txt
tqdm
beautifulsoup4
datetime
PySastrawi
requests